backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.9.0
cffi==2.0.0
channels==4.3.2
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
daphne==4.2.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
fonttools==4.62.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kiwisolver==1.5.0
matplotlib==3.10.8
msgpack==1.1.2
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
seaborn==0.13.2
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
zope.interface==8.2